pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "raven_features"
version = "0.0.4"
description = "End-to-end featurization pipeline orchestration on ClearML with YAML configs, built for the RAVEN ecosystem."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "David Levy" },
  { name = "Rushil Nagabhushan" }
]
license = { text = "Proprietary" }

dependencies = [
  "typer>=0.12.0",
  "rich>=13.0.0",
  "clearml>=1.15.0",
  "boto3>=1.28.0",
  "PyYAML>=6.0.1",
  "pydantic>=2.6.0",
  "raven @ git+https://github.com/Picture-Health/raven.git"
]

[project.urls]
Homepage = "https://github.com/Picture-Health/raven-features"

[project.optional-dependencies]
orm = ["SQLAlchemy>=2.0.0"]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "ruff>=0.5.0",
  "black>=24.3.0",
  "mypy>=1.8.0",
  "types-PyYAML>=6.0.12.12",
]

[project.scripts]
raven-features = "raven_features.cli:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
include = ["README.md", "src/raven_features/**"]

[tool.hatch.build.targets.wheel]
packages = ["src/raven_features"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "W", "UP", "B"]
ignore = ["E203"]
fix = true

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
no_implicit_optional = true
strict_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true